refactor codebase
