+++ src/io/outputWriter/XYZWriter.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "ParticleContainer.h"
+#include "models/ParticleContainer.h"
 
 
 namespace outputWriter {
